moved to client.benchmark package
